FileLogger appends to file
