+++ Logger.Tests/Logger.Tests.csproj
@@ -2,8 +2,12 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <IsPackable>false</IsPackable>
+     <IsPackable>false</IsPackable>
+     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+     <EnableNETAnalyzers>true</EnableNETAnalyzers>
+     <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
+     <LangVersion>9.0</LangVersion>
+     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
