+++ Logger/test.txt
@@ -1,0 +1,9 @@
+1/18/2021 6:26:22 PM FileLoggerTests Error the message 42
+1/18/2021 6:26:33 PM FileLoggerTests Error the message 42
+1/18/2021 6:26:40 PM FileLoggerTests Error the message 42
+1/18/2021 6:26:45 PM FileLoggerTests Error the message 42
+1/18/2021 6:27:04 PM FileLoggerTests Error the message 42
+1/18/2021 6:29:32 PM FileLoggerTests Error the message 42
+1/18/2021 6:32:30 PM FileLoggerTests Error the message 42
+1/18/2021 6:34:34 PM FileLoggerTests Error: the message 42
+1/18/2021 6:37:21 PM FileLoggerTests Error: the message 42
